drivers: rtc: max31343: fix update_bits data type
Change data parameter from uint32_t to uint8_t to
match max31343_reg_read expected input type.

Fixes: 5603551 ("drivers: rtc: add support for max31343 real-time-clock")

+++ drivers/rtc/max31343/max31343.c
@@ -85,7 +85,7 @@ int max31343_update_bits(struct max31343_dev *dev, uint8_t reg_addr,
 			 uint8_t mask, uint8_t reg_data)
 {
 	int ret;
-	uint32_t data;
+	uint8_t data;
 
 	ret = max31343_reg_read(dev, reg_addr, &data);
 	if (ret)
